Eliminate more shifting of signed integers.
To be honest, this might be a vain effort: there are many more cases of shifting signed integers throughout the codebase, and GCC considers them such a non-issue that it doesn't even have an option for producing compiler warnings when they happen. This might just be one of those cases of something that technically is against the C standard but doesn't harm portability to any significant degree (like assuming that negative signed integers are expressed in binary as two's complement).